feat: Remove WebAssembly and Add Multiplayer Support
This change eliminates WebAssembly from the project as part of a proof-of-concept to enable multiplayer gameplay across two computers over a network. All game logic has been shifted to the server, simplifying the client to focus on rendering and user input.

BREAKING CHANGE: This marks the completion of my initial goal for this project. I set out to recapture the joy of coding—those simpler times when writing code felt exciting and fulfilling. I’ve achieved that happiness and am wrapping up this journey.
https://youtu.be/jlAgHt92lqE?si=mabkplj9rDWY2MsI